+++ data/ChangelogData.lua
@@ -9,6 +9,7 @@ PER.CHANGELOG = {
 		date = buildDate ~= "" and buildDate or nil,
 		entries = {
 			"Added: Changelog window available from the options menu",
+			"Added: Changelog window available through the 'changelog' slash command",
 			"Removed: Version notice chat messages",
 			"Adapted to the latest version of Arcane Wizard: Library to ensure full compatibility"
 		}