+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
 - Added: Changelog window available from the options menu
+- Added: Changelog window available through the 'changelog' slash command
 - Removed: Version notice chat messages
 - Adapted to the latest version of Arcane Wizard: Library to ensure full compatibility